Uploaded code for analysis in paper
This is the code required for analysis and simulations in paper. The python scripts are for  simulating the various scenarios and recording summary statistics. The jupyter notebook analyze results of the simulations and plot figures.  (Notebooks also run some simple simulations which are not computationally intensive)
The "metabolic_CR_model_functions.py" and "metabolic_CR_model_class.py" define class and functions that are used by all the different codes.